+++ src/StepperControlEncoder.cpp
@@ -53,7 +53,7 @@ void StepperControlEncoder::loadPinNumbers(int channelA, int channelB, int chann
   shiftChannels();
 }
 
-void StepperControlEncoder::loadSettings(int encType, int scaling, int invert)
+void StepperControlEncoder::loadSettings(int encType, long scaling, int invert)
 {
   encoderType = encType;
   scalingFactor = scaling;
